requirements.txt
av==11.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==4.1.0
faster-whisper==1.0.1
ffmpeg-python==0.2.0
filelock==3.13.1
flatbuffers==24.3.7
fsspec==2024.2.0
future==1.0.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.16.3
packaging==24.0
protobuf==4.25.3
PyYAML==6.0.1
requests==2.31.0
sympy==1.12
tokenizers==0.15.2
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
